Prepare for multiple nonce account state versions
